Moved error message to hover tooltip & file location update
